pyproject.toml
[project]
name = "mercadobitcoin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "pymysql>=1.1.1",
    "requests>=2.32.3",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy>=2.0.39",
    "urllib3>=2.3.0",
    "celery>=5.4.0",
    "python-dotenv>=1.0.1",
    "cryptography>=44.0.2",
]
